+++ modelcraft/jobs/findwaters.py
@@ -35,15 +35,14 @@ class FindWaters(Job):
         water_model = read_structure(self._path("water.pdb"))[0]
         for water_chain in water_model:
             for water_residue in water_chain:
+                water_residue[0].element = gemmi.Element("O")
                 water_residues.append(water_residue)
         structure = self.structure.clone()
         if len(water_residues) > 0:
-            model = structure[0]
-            chain = "DUM" if self.dummy else "WAT"
-            if not model.find_chain(chain):
-                model.add_chain(chain)
+            chain = gemmi.Chain("A")
             for residue in water_residues:
-                model[chain].add_residue(residue)
+                chain.add_residue(residue)
+            structure[0].add_chain(chain, unique_name=True)
         return FindWatersResult(
             structure=structure,
             seconds=self._seconds,