+++ modelcraft/modelcraftxray.py
@@ -13,6 +13,7 @@ from .jobs.sheetbend import Sheetbend
 from .jobs.nucleofind import NucleoFind
 from .jobs.nucleofind_build import NucleoFindBuild
 from .cell import max_distortion, remove_scale, update_cell
+from .combine import combine_results
 from .pipeline import Pipeline
 from .reflections import DataItem, write_mtz
 from .structure import ModelStats, remove_residues, write_mmcif
@@ -108,8 +109,7 @@ class ModelCraftXray(Pipeline):
         if self.args.basic:
             if self.cycle == 1:
                 self.parrot()
-            self.buccaneer()
-            self.nautilus()
+            self.run_buccaneer_and_nautilus()
         else:
             if self.cycle > 1 and self.resolution < 2.3:
                 self.prune()
@@ -118,7 +118,7 @@ class ModelCraftXray(Pipeline):
                 if self.cycle > 1 or self.args.phases is None:
                     self.findwaters(dummy=True)
                 remove_residues(structure=self.current_structure, names={"HOH", "DUM"})
-            self.buccaneer()
+            self.run_buccaneer_and_nautilus()
             self.prune(chains_only=True)
             self.nucleofind()
             self.findwaters()
@@ -138,6 +138,17 @@ class ModelCraftXray(Pipeline):
         ).run(self)
         write_mmcif(self.path("current.cif"), build_result.structure)
         self.refmac(build_result.structure, cycles=10, auto_accept=True)
+            self.findwaters()
+
+    def run_buccaneer_and_nautilus(self):
+        buccaneer = self.buccaneer()
+        nautilus = self.nautilus()
+        if buccaneer is None or nautilus is None:
+            self.update_current_from_refmac_result(buccaneer or nautilus)
+        else:
+            combined = self.run_refmac(combine_results(buccaneer, nautilus), cycles=5)
+            best = min((buccaneer, nautilus, combined), key=lambda result: result.rfree)
+            self.update_current_from_refmac_result(best)
 
     def buccaneer(self):
         if not self.args.contents.proteins:
@@ -154,9 +165,10 @@ class ModelCraftXray(Pipeline):
             filter_mr=True,
             seed_mr=True,
             cycles=3 if self.cycle == 1 else 2,
+            threads=self.args.threads,
         ).run(self)
         write_mmcif(self.path("current.cif"), result.structure)
-        self.refmac(result.structure, cycles=10, auto_accept=True)
+        return self.run_refmac(result.structure, cycles=10)
 
 
     def nautilus(self):
@@ -171,7 +183,7 @@ class ModelCraftXray(Pipeline):
             structure=self.current_structure,
         ).run(self)
         write_mmcif(self.path("current.cif"), result.structure)
-        self.refmac(result.structure, cycles=5, auto_accept=True)
+        return self.run_refmac(result.structure, cycles=10)
 
     def refmac(self, structure: gemmi.Structure, cycles: int, auto_accept: bool):
         result = self.run_refmac(structure, cycles)
@@ -263,26 +275,21 @@ class ModelCraftXray(Pipeline):
     def process_cycle_output(self, result):
         _print_refmac_result(result)
         model_stats = ModelStats(result.structure)
-        stats = {"cycle": self.cycle, "residues": model_stats.residues}
-        stats["waters"] = model_stats.waters
-        stats["r_work"] = result.rwork
-        stats["r_free"] = result.rfree
+        stats = {
+            "cycle": self.cycle,
+            "residues": model_stats.residues,
+            "protein": model_stats.protein,
+            "nucleic": model_stats.nucleic,
+            "waters": model_stats.waters,
+            "r_work": result.rwork,
+            "r_free": result.rfree,
+        }
         self.report["cycles"].append(stats)
         if self.output_refmac is None or result.rwork < self.output_refmac.rwork:
             self.cycles_without_improvement = 0
             self.output_refmac = result
             write_mmcif(self.path("modelcraft.cif"), result.structure)
-            write_mtz(
-                self.path("modelcraft.mtz"),
-                [
-                    self.args.fmean,
-                    self.args.freer,
-                    result.abcd,
-                    result.fphi_best,
-                    result.fphi_diff,
-                    result.fphi_calc,
-                ],
-            )
+            result.mtz.write_to_file(self.path("modelcraft.mtz"))
             self.report["final"] = stats
         else:
             self.cycles_without_improvement += 1
@@ -319,7 +326,10 @@ class ModelCraftXray(Pipeline):
 
 def _print_refmac_result(result):
     model_stats = ModelStats(result.structure)
-    print(f"\nResidues: {model_stats.residues:6d}", flush=True)
-    print(f"Waters:   {model_stats.waters:6d}", flush=True)
-    print(f"R-work:   {result.rwork:6.4f}", flush=True)
+    print("")
+    print(f"Residues: {model_stats.residues:6d}")
+    print(f"Protein:  {model_stats.protein:6d}")
+    print(f"Nucleic:  {model_stats.nucleic:6d}")
+    print(f"Waters:   {model_stats.waters:6d}")
+    print(f"R-work:   {result.rwork:6.4f}")
     print(f"R-free:   {result.rfree:6.4f}", flush=True)